+++ airline-web/app/controllers/AllianceMissionApplication.scala
@@ -3,7 +3,7 @@ package controllers
 import com.patson.AllianceMissionSimulation
 import com.patson.data.{AllianceMissionSource, AllianceSource, CycleSource}
 import com.patson.model._
-import com.patson.model.alliance.AllianceMissionStatus
+import com.patson.model.alliance.{AllianceMission, AllianceMissionReward, AllianceMissionStatus}
 import controllers.AuthenticationObject.AuthenticatedAirline
 import play.api.mvc._
 import play.api.libs.json._
@@ -35,13 +35,7 @@ class AllianceMissionApplication @Inject()(cc: ControllerComponents) extends Abs
                   mission.status = AllianceMissionStatus.SELECTED
                   AllianceMissionSource.updateAllianceMission(mission)
 
-                  val missions = AllianceMissionSource.loadAllianceMissionsAfterCutoff(allianceMember.allianceId, cycle - AllianceMissionSimulation.MISSION_DURATION)
-
-                  var result = Json.obj("missionCandidates" -> missions.filter(entry => entry.status == AllianceMissionStatus.CANDIDATE || entry.status == AllianceMissionStatus.SELECTED).sortBy(_.id))
-                  missions.find(_.status == AllianceMissionStatus.SELECTED).foreach { selectedMission =>
-                    result = result + ("selectedMission" -> Json.toJson(selectedMission))
-                  }
-                  Ok(result)
+                  Ok(AllianceMissionUtil.buildMissionJson(allianceMember.allianceId))
                 }
               }
         } else {
